# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/recommendation": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List Recommendations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommendation"
                ],
                "summary": "List Recommendations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Recommendated By",
                        "name": "recommendated_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Priority",
                        "name": "priority",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_sync.GetAllRecommendationReq"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create Recommendation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommendation"
                ],
                "summary": "Create Recommendation",
                "parameters": [
                    {
                        "description": "body for RecommendationCreate",
                        "name": "recommendation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_sync.RecommendationCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recommendation created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/recommendation/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Recommendation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommendation"
                ],
                "summary": "Get Recommendation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recommendation ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_sync.RecommendationRes"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Recommendation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommendation"
                ],
                "summary": "Update Recommendation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recommendation ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "body for RecommendationUpdate",
                        "name": "recommendation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_sync.RecommendationUpdateBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recommendation updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete Recommendation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recommendation"
                ],
                "summary": "Delete Recommendation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recommendation ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recommendation deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "health_sync.Filter": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                }
            }
        },
...
        "health_sync.GetAllRecommendationReq": {
            "type": "object",
            "properties": {
                "filter": {
                    "$ref": "#/definitions/health_sync.Filter"
                },
                "priority": {
                    "type": "integer"
                },
                "recommendated_by": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "health_sync.RecommendationCreate": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "recommendated_by": {
                    "type": "string"
                },
                "recommendation_type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "health_sync.RecommendationRes": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "recommendated_by": {
                    "type": "string"
                },
                "recommendation_type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "health_sync.RecommendationUpdateBody": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "recommendation_type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}